bookingsys/new_backend/requirements.txt
motor==3.6.0
pymongo==4.9.2
certifi
fastapi
uvicorn
sqlalchemy
jinja2
email-validator
pydantic-settings


